+++ core/src/com/unciv/ui/screens/civilopediascreen/CivilopediaCategories.kt
@@ -16,105 +16,103 @@ import com.unciv.models.ruleset.unit.UnitType as BaseUnitType
  */
 enum class CivilopediaCategories (
     val label: String,
-    val hide: Boolean,      // Omitted on CivilopediaScreen
     val getImage: ((name: String, size: Float) -> Actor?)?,
     val binding: KeyboardBinding,
     val headerIcon: String,
     val getCategoryIterator: (ruleset: Ruleset, tutorialController: TutorialController) -> Collection<ICivilopediaText>
 ) {
-    Building ("Buildings", false,
+    Building ("Buildings",
         CivilopediaImageGetters.construction,
         KeyboardBinding.PediaBuildings,
         "OtherIcons/Cities",
         { ruleset, _ -> ruleset.buildings.values.filter { !it.isAnyWonder() } }
     ),
-    Wonder ("Wonders", false,
+    Wonder ("Wonders",
         CivilopediaImageGetters.construction,
         KeyboardBinding.PediaWonders,
         "OtherIcons/Wonders",
         { ruleset, _ -> ruleset.buildings.values.filter { it.isAnyWonder() } }
     ),
-    Resource ("Resources", false,
+    Resource ("Resources",
         CivilopediaImageGetters.resource,
         KeyboardBinding.PediaResources,
         "OtherIcons/Resources",
         { ruleset, _ -> ruleset.tileResources.values }
     ),
-    Terrain ("Terrains", false,
+    Terrain ("Terrains",
         CivilopediaImageGetters.terrain,
         KeyboardBinding.PediaTerrains,
         "OtherIcons/Terrains",
         { ruleset, _ -> ruleset.terrains.values }
     ),
-    Improvement ("Tile Improvements", false,
+    Improvement ("Tile Improvements",
         CivilopediaImageGetters.improvement,
         KeyboardBinding.PediaImprovements,
         "OtherIcons/Improvements",
         { ruleset, _ -> ruleset.tileImprovements.values }
     ),
-    Unit ("Units", false,
+    Unit ("Units",
         CivilopediaImageGetters.construction,
         KeyboardBinding.PediaUnits,
         "OtherIcons/Shield",
         { ruleset, _ -> ruleset.units.values }
     ),
-    UnitType ("Unit types", false,
+    UnitType ("Unit types",
         CivilopediaImageGetters.unitType,
         KeyboardBinding.PediaUnitTypes,
         "UnitTypeIcons/UnitTypes",
         { ruleset, _ -> BaseUnitType.getCivilopediaIterator(ruleset) }
     ),
-    Nation ("Nations", false,
+    Nation ("Nations",
         CivilopediaImageGetters.nation,
         KeyboardBinding.PediaNations,
         "OtherIcons/Nations",
         { ruleset, _ -> ruleset.nations.values.filter { !it.isSpectator } }
     ),
-    Technology ("Technologies", false,
+    Technology ("Technologies",
         CivilopediaImageGetters.technology,
         KeyboardBinding.PediaTechnologies,
         "TechIcons/Philosophy",
         { ruleset, _ -> ruleset.technologies.values }
     ),
-    Promotion ("Promotions", false,
+    Promotion ("Promotions",
         CivilopediaImageGetters.promotion,
         KeyboardBinding.PediaPromotions,
         "UnitPromotionIcons/Mobility",
         { ruleset, _ -> ruleset.unitPromotions.values }
     ),
-    Policy ("Policies", false,
+    Policy ("Policies",
         CivilopediaImageGetters.policy,
         KeyboardBinding.PediaPolicies,
         "PolicyIcons/Constitution",
         { ruleset, _ -> ruleset.policies.values }
     ),
     Belief("Religions and Beliefs",
-        hide = false, // Skipping this page when showReligionInCivilopedia returns false is hardcoded
         CivilopediaImageGetters.belief,
         KeyboardBinding.PediaBeliefs,
         "ReligionIcons/Religion",
         { ruleset, _ -> (ruleset.beliefs.values.asSequence() +
             BaseBelief.getCivilopediaReligionEntry(ruleset)).toList() }
     ),
-    Tutorial ("Tutorials", false,
+    Tutorial ("Tutorials",
         getImage = null,
         KeyboardBinding.PediaTutorials,
         "OtherIcons/ExclamationMark",
         { _, tutorialController -> tutorialController.getCivilopediaTutorials() }
     ),
-    Difficulty ("Difficulty levels", false,
+    Difficulty ("Difficulty levels",
         getImage = null,
         KeyboardBinding.PediaDifficulties,
         "OtherIcons/Quickstart",
         { ruleset, _ -> ruleset.difficulties.values }
     ),
-    Era ("Eras", false,
+    Era ("Eras",
         getImage = null,
         KeyboardBinding.PediaEras,
         "OtherIcons/Tyrannosaurus",
         { ruleset, _ -> ruleset.eras.values }
     ),
-    Speed ("Speeds", false,
+    Speed ("Speeds",
         getImage = null,
         KeyboardBinding.PediaSpeeds,
         "OtherIcons/Timer",
@@ -125,6 +123,5 @@ enum class CivilopediaCategories (
         fun fromLink(name: String): CivilopediaCategories? =
             values().firstOrNull { it.name == name }
             ?: values().firstOrNull { it.label == name }
-
     }
 }